QCamera2: HAL: Restart the preview channel
Issue:
In some of the 3rd party Application's, Torch mode is set
in preview stopped state with default dimensions and
also recording hint is set to false.
so, we add a preview channel thinking it.s a flash case and
then App is setting the actual preview dimensions which we
don't honor since the preview channel is already there and
due to this mismatch in dimensions, garbled preview is seen.

Fix:
Delete the Preview Channel if it.s already there and add a
new preview channel with updated dimensions & format.

Change-Id: Ia1bd0d30a7939b8bdfd5d88fe100f696dfb4a7c6