QCamera2: Fix choppy video preview in low-light after recording video
The recording hint must be disabled in order to fix a bug in Google Camera.
However, the anti-shake algorithm uses the recording hint value to determine
whether camcorder mode is enabled.

Cache the userspace-set recording hint value and use that for anti-shake instead
to ensure that the proper exposure mode setting is selected.